Complete data source interface separation
Removes monolithic LocalDataSource and RemoteDataSource interfaces in favor of functional separation (Chat, Message, Sync). Extracts error types into separate files and updates DI bindings.
