+++ app/src/main/java/timur/gilfanov/messenger/di/DataSourceModule.kt
@@ -6,12 +6,10 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 import timur.gilfanov.messenger.data.source.local.LocalChatDataSource
-import timur.gilfanov.messenger.data.source.local.LocalDataSource
 import timur.gilfanov.messenger.data.source.local.LocalDataSourceFake
 import timur.gilfanov.messenger.data.source.local.LocalMessageDataSource
 import timur.gilfanov.messenger.data.source.local.LocalSyncDataSource
 import timur.gilfanov.messenger.data.source.remote.RemoteChatDataSource
-import timur.gilfanov.messenger.data.source.remote.RemoteDataSource
 import timur.gilfanov.messenger.data.source.remote.RemoteDataSourceFake
 import timur.gilfanov.messenger.data.source.remote.RemoteMessageDataSource
 import timur.gilfanov.messenger.data.source.remote.RemoteSyncDataSource
@@ -39,11 +37,6 @@ abstract class DataSourceModule {
         localDataSourceFake: LocalDataSourceFake,
     ): LocalSyncDataSource
 
-    // Keep the original LocalDataSource binding for backward compatibility
-    @Binds
-    @Singleton
-    abstract fun bindLocalDataSource(localDataSourceFake: LocalDataSourceFake): LocalDataSource
-
     // Remote data source bindings - all interfaces bound to the same implementation
     @Binds
     @Singleton
@@ -62,9 +55,4 @@ abstract class DataSourceModule {
     abstract fun bindRemoteSyncDataSource(
         remoteDataSourceFake: RemoteDataSourceFake,
     ): RemoteSyncDataSource
-
-    // Keep the original RemoteDataSource binding for backward compatibility
-    @Binds
-    @Singleton
-    abstract fun bindRemoteDataSource(remoteDataSourceFake: RemoteDataSourceFake): RemoteDataSource
 }
